feat(GlobalProvider): Created a new global provider to control loading
