+++ client/src/App.jsx
@@ -2,6 +2,7 @@ import PropTypes from 'prop-types';
 import { BrowserRouter as Router, Routes, Route, Navigate } from 'react-router-dom';
 import { AuthProvider, useAuth } from './utils/AuthContext';
 import { ThemeProvider } from '../src/pages/Settings/ThemeContext';
+import { GlobalProvider } from './pages/Loading/GlobalProvider';
 import { ReadingContextProvider } from './context/ReadingContext';
 import Layout from './components/AppLayout/Layout';
 
@@ -9,9 +10,11 @@ const App = () => {
     return (
         <AuthProvider>
             <ThemeProvider>
-                <Router>
-                    <MainRoutes />
-                </Router>
+                <GlobalProvider>
+                    <Router>
+                        <MainRoutes />
+                    </Router>
+                </GlobalProvider>
             </ThemeProvider>
         </AuthProvider>
     );
